backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
requests
flask
flask-cors
mysql-connector-python
python-dotenv
